Add people list and form
These display a list of people from the API. A person can be created through a
form alongside the list. People from the list can be loaded into the form for
updating.

The views are styled using Bootstrap and are accessible on all screens sizes.
Axe browser tool reports no accessibility issues.

The view is set as the home page.