Surface API response errors as error summary
Whilst this shouldn't happen as the frontend validation is the same as the
backend validation, it might be useful to surface error responses from the API
on the form.

Currently, this is done as an error summary. It might be a nice refinement to
map these to the input fields to be more in-keeping with the other validation.

This also merges the create and update methods to simplify the component model.